vc_hunter_v2/requirements.txt
beautifulsoup4
requests
tqdm
openai
python-dotenv
matplotlib
scikit-learn
openai
python-dotenv     # optional, but good if you ever use .env files
tiktoken
